Commit for Vlad.File my_module_my_each_spec.rb dont work.
